Created HeartbeatThreadHelper
