+++ core/src/test/scala/unit/kafka/zk/ZooKeeperTestHarness.scala
@@ -30,7 +30,7 @@ import org.junit.experimental.categories.Category
 import scala.collection.Set
 import scala.collection.JavaConverters._
 import org.apache.kafka.clients.producer.KafkaProducer
-import org.apache.kafka.clients.consumer.internals.AbstractCoordinator
+import org.apache.kafka.clients.consumer.internals.HeartbeatThread
 import kafka.controller.ControllerEventManager
 import org.apache.kafka.common.utils.Time
 import org.apache.zookeeper.{WatchedEvent, Watcher, ZooKeeper}
@@ -91,7 +91,7 @@ object ZooKeeperTestHarness {
   // which reset static JAAS configuration.
   val unexpectedThreadNames = Set(ControllerEventManager.ControllerEventThreadName,
                                   KafkaProducer.NETWORK_THREAD_PREFIX,
-                                  AbstractCoordinator.HEARTBEAT_THREAD_PREFIX,
+                                  HeartbeatThread.HEARTBEAT_THREAD_PREFIX,
                                   ZkClientEventThreadPrefix)
 
   /**
